requirements.txt
django>=3.2
psycopg[binary]
python-dotenv
channels
daphne
djangorestframework
web3
python-json-logger==2.0.7

